// src/dave/protocol/v1/enums.ts
export enum LeafNodeSource {
  RESERVED = 0,
  KEY_PACKAGE = 1,
  UPDATE = 2,
  COMMIT = 3,
}

export enum ProposalOperationType {
  APPEND = 0,
  REVOKE = 1,
}

export enum ExtensionType {
  RESERVED = 0,

  APPLICATION_ID = 1,
  RATCHET_TREE = 2,
  REQUIRED_CAPABILITIES = 3,
  EXTERNAL_PUB = 4,
  EXTERNAL_SENDERS = 5,

  GREASE = 0x0A0A,
  GREASE_1 = 0x1A1A,
  GREASE_2 = 0x2A2A,
  GREASE_3 = 0x3A3A,
  GREASE_4 = 0x4A4A,
  GREASE_5 = 0x5A5A,
  GREASE_6 = 0x6A6A,
  GREASE_7 = 0x7A7A,
  GREASE_8 = 0x8A8A,
  GREASE_9 = 0x9A9A,
  GREASE_10 = 0xAAAA,
  GREASE_11 = 0xBABA,
  GREASE_12 = 0xCACA,
  GREASE_13 = 0xDADA,
  GREASE_14 = 0xEAEA,
}

export enum CredentialType {
  RESERVED = 0,

  BASIC = 1,
  X509 = 2,

  GREASE = 0x0A0A,
  GREASE_1 = 0x1A1A,
  GREASE_2 = 0x2A2A,
  GREASE_3 = 0x3A3A,
  GREASE_4 = 0x4A4A,
  GREASE_5 = 0x5A5A,
  GREASE_6 = 0x6A6A,
  GREASE_7 = 0x7A7A,
  GREASE_8 = 0x8A8A,
  GREASE_9 = 0x9A9A,
  GREASE_10 = 0xAAAA,
  GREASE_11 = 0xBABA,
  GREASE_12 = 0xCACA,
  GREASE_13 = 0xDADA,
  GREASE_14 = 0xEAEA,
}

export enum ProposalType {
  RESERVED = 0,

  ADD = 1,
  UPDATE = 2,
  REMOVE = 3,
  PSK = 4,
  REINIT = 5,
  EXTERNAL_INIT = 6,
  GROUP_CONTEXT_EXTENSIONS = 7,

  GREASE = 0x0A0A,
  GREASE_1 = 0x1A1A,
  GREASE_2 = 0x2A2A,
  GREASE_3 = 0x3A3A,
  GREASE_4 = 0x4A4A,
  GREASE_5 = 0x5A5A,
  GREASE_6 = 0x6A6A,
  GREASE_7 = 0x7A7A,
  GREASE_8 = 0x8A8A,
  GREASE_9 = 0x9A9A,
  GREASE_10 = 0xAAAA,
  GREASE_11 = 0xBABA,
  GREASE_12 = 0xCACA,
  GREASE_13 = 0xDADA,
  GREASE_14 = 0xEAEA,
}

export enum PSKType {
  RESERVED = 0,

  EXTERNAL = 1,
  RESUMPTION = 2,
}

export enum ResumptionPSKUsage {
  RESERVED = 0,

  APPLICATION = 1,
  REINIT = 2,
  BRANCH = 3,
}

export enum ProtocolVersion {
  RESERVED = 0,

  MLS_10 = 1,
}

export enum ContentType {
  RESERVED = 0,

  APPLICATION = 1,
  PROPOSAL = 2,
  COMMIT = 3,
}

export enum WireFormats {
  RESERVED = 0,

  MLS_PUBLIC_MESSAGE = 1,
  MLS_PRIVATE_MESSAGE = 2,
  MLS_WELCOME = 3,
  MLS_GROUP_INFO = 4,
  MLS_KEY_PACKAGE = 5,
}

export enum SenderType {
  RESERVED = 0,

  MEMBER = 1,
  EXTERNAL = 2,
  NEW_MEMBER_PROPOSAL = 3,
  NEW_MEMBER_COMMIT = 4,
}

export enum CipherSuite {
  RESERVED = 0,

  MLS_128_DHKEMX25519_AES128GCM_SHA256_ED25519 = 1,
  MLS_128_DHKEMP256_AES128GCM_SHA256_P256 = 2,
  MLS_128_DHKEMX25519_CHACHA20POLY1305_SHA256_ED25519 = 3,
  MLS_256_DHKEMX448_AES256GCM_SHA512_ED448 = 4,
  MLS_256_DHKEMP521_AES256GCM_SHA512_P521 = 5,
  MLS_256_DHKEMX448_CHACHA20POLY1305_SHA512_ED448 = 6,
  MLS_256_DHKEMP384_AES256GCM_SHA384_P384 = 7,

  GREASE = 0x0A0A,
  GREASE_1 = 0x1A1A,
  GREASE_2 = 0x2A2A,
  GREASE_3 = 0x3A3A,
  GREASE_4 = 0x4A4A,
  GREASE_5 = 0x5A5A,
  GREASE_6 = 0x6A6A,
  GREASE_7 = 0x7A7A,
  GREASE_8 = 0x8A8A,
  GREASE_9 = 0x9A9A,
  GREASE_10 = 0xAAAA,
  GREASE_11 = 0xBABA,
  GREASE_12 = 0xCACA,
  GREASE_13 = 0xDADA,
  GREASE_14 = 0xEAEA,
}

export enum ProposalOrRefType {
  RESERVED = 0,

  PROPOSAL = 1,
  REFERENCE = 2,
}